Use Int64 for all token count fields (CSSLP integer safety rule)
Token counts can accumulate to tens of millions across a weekly window.
Explicitly use Int64 throughout rather than relying on Int being 64-bit
on the current platform, per the CSSLP checklist in CLAUDE.md.

Changes: UsageTokens, ContextWindowMetrics, WeeklyUsageMetrics,
BillingWindowMetrics, ModelLimits, both calculators, WeeklyUsageSection.
36 tests still passing.

+++ ClaudeContextMeter/ClaudeContextMeter/Models/SessionRecord.swift
@@ -20,10 +20,10 @@ struct MessageContent: Codable {
 }
 
 struct UsageTokens: Codable {
-    let inputTokens: Int
-    let cacheCreationInputTokens: Int?
-    let cacheReadInputTokens: Int?
-    let outputTokens: Int
+    let inputTokens: Int64
+    let cacheCreationInputTokens: Int64?
+    let cacheReadInputTokens: Int64?
+    let outputTokens: Int64
 
     enum CodingKeys: String, CodingKey {
         case inputTokens = "input_tokens"
@@ -33,7 +33,7 @@ struct UsageTokens: Codable {
     }
 
     /// Total tokens consumed in this turn (input + cache + output).
-    var totalTokens: Int {
+    var totalTokens: Int64 {
         inputTokens + (cacheCreationInputTokens ?? 0) + (cacheReadInputTokens ?? 0) + outputTokens
     }
 }
